% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifacts.R
\name{assemble_facs_panels}
\alias{assemble_facs_panels}
\title{Assemble editable ggplots on a dynamic Plotgardener page}
\usage{
assemble_facs_panels(plots, ncol = NULL, layout_options = list())
}
\arguments{
\item{plots}{A named list of ggplot objects.}

\item{ncol}{Number of columns.}

\item{layout_options}{Optional Plotgardener sizing settings.}
}
\description{
Assemble editable ggplots on a dynamic Plotgardener page
}
