% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph3.R
\name{plot_ph3_overall}
\alias{plot_ph3_overall}
\title{Plot the overall pH3-positive percentage}
\usage{
plot_ph3_overall(analysis, appearance = NULL, appearance_file = NULL)
}
\arguments{
\item{analysis}{A quantified PH3 analysis.}

\item{appearance}{Optional appearance overrides.}

\item{appearance_file}{Optional appearance YAML.}
}
\description{
Plot the overall pH3-positive percentage
}
