% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifacts.R
\name{save_facs_figure_bundle}
\alias{save_facs_figure_bundle}
\alias{read_facs_figure_bundle}
\title{Save or read a FACS figure bundle}
\usage{
save_facs_figure_bundle(bundle, path, overwrite = FALSE)

read_facs_figure_bundle(path)
}
\arguments{
\item{bundle}{A \code{facs_figure_bundle} object.}

\item{path}{Explicit RDS path.}

\item{overwrite}{Whether an existing file may be replaced.}
}
\description{
Save or read a FACS figure bundle
}
